Mostly implement /part and buffer hiding
Buffers are (currently) never deleted, but parting from a channel hides
the buffer until new messages are received in that channel. In the case
of /part, this shouldn't happen (the server won't send us messages for
channels we're not in). In the case of /hide, this could happen, in
which case the buffer is reactivated.

Currently there is a race-condition between the server sending the part
to be reactivated immediately. To get around this, some smart detection
will be required when accessing buffers.