perf: specify __slots__ for Message
Since you may be storing lots of Messages in lots of Buffers in RAM, it
makes sense to make the Message class as efficient as possible by
explicitly declaring its attributes and avoiding allocating space for a
dcit for each message. Also comes with the nice side-effect of making
message attribute lookups faster.